fix(test): split spartan_pytest tests into per-module mirror dir (PS-202/204 + STX-TQ002)
CI flagged three audit violations on tests/scitex_agent_container/
cli_pkg/test_spartan_pytest_cmds.py — orphan test file (no matching
src/.../spartan_pytest_cmds.py), no mirror dir under
src/scitex_agent_container/cli_pkg/spartan_pytest/, and three
STX-TQ002 AAA-marker combinations (`# Arrange / Act`,
`# Act / Assert`).

Fix: split the monolithic test file into per-module test files
matching the src package layout (PS-205 double-underscore
convention for private modules):

- test__render.py    (5 tests) -> _render.py
- test__summary.py   (13 tests) -> _summary.py
- test__ssh.py       (2 tests)  -> _ssh.py
- test__run_cmd.py   (4 tests)  -> _run_cmd.py

The three TQ002 violations are fixed in-place:
- test_render_invokes_pytest_with_no_cov_flag: extracted kwargs
  into Arrange so Act stands alone.
- test_split_repo_at_branch_rejects_missing_at: switched from
  with pytest.raises to the try/except-into-raised pattern
  used elsewhere in the suite (stx-allow: test-capture).
- test_default_reservation_is_operator_sapphire_pool: pulled the
  expected constant into Arrange so Act is the actual lookup.

Verified locally: pytest tests/.../spartan_pytest/ -q --no-cov
-> 23 passed, 1 skipped (live Spartan round-trip behind
$SAC_SPARTAN_HOST as designed).
